Create tkinter boiler coding
This is a boiler software which used to determine the boiler efficiency by using Indirect method for the Solid, Liquid and Gas Fuels. The user can also key in all the need information to analyze the performance of the boiler. The obtained results can be copied to an Notepad for future use. The resolution is adjusted to the